apps/api/pyproject.toml
[project]
name = "tigerhacks25-backend"
version = "1.0.0"
description = "Backend API for Planet Evolution Game"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.9.0",
    "python-dotenv>=1.0.0",
]
